Add tool mocks
This allows a chained execution to use the output of a previous tool call, without actually causing the side-effects of calling the tool. You can get a mocked tool call result back to the model in the middle of a chain to get it going in a certain direction - usually toward the next tool call you expect it to make.